+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(getting-started): implement Getting Started feature with documentation and UI integration. (`266123a`)
 - fix(docs): update internalDocumentIDStore in docsSearchIndex.json. (`249a3b5`)
 - fix(docs): update docsSearchIndex.json with new internal document IDs. (`bd47ea4`)
 - feat(scripting): enhance module import handling and testing. (`87d263e`)